requirements.txt
flask
flask_cors
gunicorn
numpy
pandas
easyocr
Pillow
ultralytics
opencv-python
python-dotenv